[Workload] split data and query workload
