+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/workload/interfaces/IDataWorkLoad.java
@@ -1,0 +1,22 @@
+package cn.edu.tsinghua.iotdb.benchmark.workload.interfaces;
+
+import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
+import cn.edu.tsinghua.iotdb.benchmark.exception.WorkloadException;
+
+/** interface of data workload */
+public interface IDataWorkLoad extends IWorkLoad {
+  /**
+   * Insert one batch into database NOTICE: every row contains data from all sensors
+   *
+   * @return
+   * @throws WorkloadException
+   */
+  Batch getOneBatch() throws WorkloadException;
+
+  /**
+   * Get Batch Number
+   *
+   * @return
+   */
+  long getBatchNumber();
+}
