+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/workload/interfaces/IQueryWorkLoad.java
@@ -1,55 +1,10 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
 package cn.edu.tsinghua.iotdb.benchmark.workload.interfaces;
 
 import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
 import cn.edu.tsinghua.iotdb.benchmark.exception.WorkloadException;
-import cn.edu.tsinghua.iotdb.benchmark.schema.schemaImpl.DeviceSchema;
 import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.*;
 
-/** IGenerateWorkload is a workload using for data which is generated by benchmark */
-public interface IGenerateDataWorkload extends IWorkLoad {
-
-  /**
-   * Insert one batch into database NOTICE: every row contains data from all sensors
-   *
-   * @param deviceSchema
-   * @param loopIndex
-   * @return
-   * @throws WorkloadException
-   */
-  Batch getOneBatch(DeviceSchema deviceSchema, long loopIndex) throws WorkloadException;
-
-  /**
-   * Insert one batch into database NOTICE: every row contains data from sensor which index is
-   * colIndex
-   *
-   * @param deviceSchema
-   * @param loopIndex
-   * @param colIndex
-   * @return
-   * @throws WorkloadException
-   */
-  Batch getOneBatch(DeviceSchema deviceSchema, long loopIndex, int colIndex)
-      throws WorkloadException;
-
+public interface IQueryWorkLoad extends IWorkLoad {
   /**
    * Get precise query Eg. select v1... from data where time = ? and device in ?
    *
@@ -116,6 +71,15 @@ public interface IGenerateDataWorkload extends IWorkLoad {
    * @throws WorkloadException
    */
   LatestPointQuery getLatestPointQuery() throws WorkloadException;
+
+  /**
+   * Return a verified Query
+   *
+   * @param batch have the data to check
+   * @return
+   * @throws WorkloadException
+   */
+  VerificationQuery getVerifiedQuery(Batch batch) throws WorkloadException;
 
   /**
    * Get device query, Eg. select time, v1... where device = ?
