--- core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/workload/interfaces/IRealDataWorkload.java
@@ -1,50 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package cn.edu.tsinghua.iotdb.benchmark.workload.interfaces;
-
-import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
-import cn.edu.tsinghua.iotdb.benchmark.exception.WorkloadException;
-import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.VerificationQuery;
-
-/** IRealDataWorkload is a workload using for real data */
-public interface IRealDataWorkload extends IWorkLoad {
-  /**
-   * Return a batch from real data return null if there is no data
-   *
-   * @return
-   * @throws WorkloadException
-   */
-  Batch getOneBatch() throws WorkloadException;
-
-  /**
-   * Return a verified Query
-   *
-   * @return
-   * @throws WorkloadException
-   */
-  VerificationQuery getVerifiedQuery() throws WorkloadException;
-
-  /**
-   * Get batch number of workload
-   *
-   * @return
-   */
-  int getBatchNumber();
-}
